1、修改配置流程
