+++ lib/http/xhrAdapter.ts
@@ -1,12 +1,10 @@
-import { HtptClientResponseOption, HttpClientRequestOption, IHttpAdapter } from "./adapterType";
+import { HtptClientResponseOption, HttpClientRequestOption, IHttpAdapter } from "./adapter";
 
 export default class XHRAdapter implements IHttpAdapter {
 
   send<T>(options: HttpClientRequestOption): Promise<HtptClientResponseOption<T>> {
     return new Promise<HtptClientResponseOption<T>>((resolve, reject) => {
-      reject({
-        message: '1111'
-      })
+      reject({})
     })
   }
