+++ tsconfig.json
@@ -1,7 +1,8 @@
 {
     "compilerOptions": {
+        "target": "ES5",
         "module": "ESNext",
-        "lib": ["dom", "es2015"],
+        "lib": ["DOM", "ES2015"],
         "types": ["node"],
         "strict": true,
         "noEmit": true
